Clean up OWC scripts
