+++ OWC_writeup/.gitignore
@@ -3,3 +3,4 @@
 *.bbl
 *.gz
 *.aux
+*.p
