+++ OWC_writeup/scripts/Ca_performance_table.csv
@@ -1,0 +1,117 @@
+inv_TSR,Ca
+-1000000,0.092282521
+-100,0.092282521
+-50,0.092282521
+-20,0.092282521
+-10,0.092282521
+0,0.092282521
+0.012668019,0.092282521
+0.03426608,0.108898362
+0.045142045,0.166429569
+0.057553153,0.220299948
+0.071173399,0.281008945
+0.085301268,0.337880945
+0.097366303,0.389410575
+0.118841776,0.459040664
+0.129754941,0.518464217
+0.139778827,0.578526192
+0.158789719,0.629890894
+0.169510357,0.672544117
+0.179040696,0.723541827
+0.19001597,0.778887736
+0.20320735,0.830091701
+0.217075432,0.892335842
+0.235373075,0.945980007
+0.243281766,0.994584051
+0.261535771,1.05696206
+0.276437612,1.106830028
+0.28738014,1.161446822
+0.309746514,1.228899216
+0.314496222,1.290898399
+0.330525042,1.359077526
+0.346010286,1.408090836
+0.35791277,1.465159246
+0.39363278,1.562243494
+0.426007626,1.655088735
+0.432751135,1.697592835
+0.448256149,1.759255085
+0.472991905,1.811679549
+0.49238604,1.859073611
+0.506538082,1.914558511
+0.522258542,1.982617847
+0.541141927,2.036904363
+0.559711436,2.087541557
+0.587280356,2.145242742
+0.603243099,2.194618208
+0.623148907,2.255209854
+0.643251111,2.300621402
+0.66566147,2.344023427
+0.680382139,2.396784052
+0.702029778,2.445120045
+0.72676211,2.495353764
+0.75149701,2.547230542
+0.778443456,2.59005665
+0.802495913,2.643459039
+0.830306138,2.691239423
+0.859654739,2.737428015
+0.890547726,2.785870789
+0.921439344,2.833437265
+0.953872962,2.88173104
+0.987842604,2.92692827
+1.021805398,2.967744011
+1.055763211,3.005373215
+1.089717911,3.041010834
+1.123670121,3.075055183
+1.157624821,3.110692802
+1.191565203,3.137169126
+1.225508697,3.165637035
+1.259449701,3.192511676
+1.293384479,3.215403146
+1.327313966,3.23490892
+1.361237849,3.250829839
+1.395167335,3.270335613
+1.429090907,3.286057374
+1.463007631,3.297397646
+1.496926223,3.30993287
+1.530844503,3.322268935
+1.564760293,3.333011731
+1.598674838,3.342957894
+1.632587515,3.351709104
+1.666499881,3.360261157
+1.700410691,3.367817416
+1.734322745,3.37617031
+1.768229819,3.381336666
+1.802136271,3.386104705
+1.836043345,3.391271062
+1.869948551,3.395242467
+1.903858116,3.402002092
+1.937757408,3.402189484
+1.971661992,3.405762572
+2.005578093,3.406670453
+2.039467113,3.410319686
+2.073365782,3.410108761
+2.107270366,3.413681849
+2.14117246,3.415661668
+2.175075487,3.418238963
+2.208974156,3.418028038
+2.242872826,3.417817113
+2.276771495,3.417606188
+2.310670165,3.417395263
+2.344568834,3.417184338
+2.378467504,3.416973413
+2.412366173,3.416762488
+2.446264843,3.416551563
+2.480163512,3.416340638
+2.503649469,3.408526823
+2.509371154,3.408526823
+2.65,3.408526823
+2.75,3.408526823
+3,3.408526823
+4,3.408526823
+5,3.408526823
+7,3.408526823
+10,3.408526823
+20,3.408526823
+50,3.408526823
+100,3.408526823
+1000000,3.408526823
