+++ OWC_writeup/scripts/Ct_performance_table.csv
@@ -1,0 +1,117 @@
+inv_TSR,Ct
+-1000000,0.375722776
+-100,0.375722776
+-10,0.375722776
+-1,0.375722776
+0,0.375722776
+0.026828855,0.375722776
+0.061012163,0.375722776
+0.095195471,0.374275093
+0.129390802,0.376102447
+0.163601319,0.382066689
+0.19783525,0.394408635
+0.232095125,0.413817765
+0.266361327,0.434950599
+0.300659802,0.464874322
+0.337110351,0.494891355
+0.371665655,0.522413413
+0.403564085,0.557058674
+0.436495224,0.595505788
+0.465399555,0.637675307
+0.514548355,0.733645009
+0.544918292,0.775246014
+0.571192363,0.808599425
+0.59270955,0.858889048
+0.618927127,0.898233279
+0.642745995,0.942662941
+0.66620479,0.98307891
+0.69188684,1.0176516
+0.739707487,1.109300279
+0.76472745,1.151661245
+0.789742192,1.192600155
+0.813204716,1.234031877
+0.8366742,1.277359673
+0.861685462,1.317350546
+0.886707165,1.36018553
+0.911725387,1.402072477
+0.936747091,1.44490746
+0.960198676,1.483359638
+0.985222229,1.526698737
+1.013346631,1.568074078
+1.041475672,1.610713468
+1.06804186,1.650948891
+1.093049642,1.689991727
+1.121175783,1.731841086
+1.152420632,1.776591773
+1.182099998,1.818222198
+1.208656035,1.855692513
+1.238318497,1.892718255
+1.272654306,1.932811829
+1.306992646,1.973594884
+1.341322127,2.011964754
+1.375644648,2.04843855
+1.409964004,2.084050494
+1.444286525,2.120524291
+1.478596389,2.15355068
+1.512902457,2.185542846
+1.547211689,2.218396865
+1.581516492,2.250044291
+1.615813701,2.279623273
+1.650105848,2.307823292
+1.684399893,2.336540422
+1.718689508,2.364050959
+1.752971531,2.389493052
+1.78725735,2.415969367
+1.82154127,2.441928572
+1.855813168,2.464612739
+1.890090128,2.488675869
+1.924355698,2.509636333
+1.958615573,2.529045463
+1.992872284,2.547592742
+2.027130893,2.566657132
+2.061392033,2.586411003
+2.095648111,2.604785911
+2.129901658,2.622471338
+2.16414951,2.638605431
+2.1983923,2.653360562
+2.232637621,2.668805174
+2.266882941,2.684249786
+2.301128262,2.699694399
+2.335369154,2.713932418
+2.369611943,2.728687549
+2.403853467,2.743097939
+2.438098788,2.758542551
+2.467671948,2.771242526
+2.6,2.771242526
+3,2.771242526
+10,2.771242526
+50,2.771242526
+100,2.771242526
+1000000,2.771242526
+2.039467113,3.410319686
+2.073365782,3.410108761
+2.107270366,3.413681849
+2.14117246,3.415661668
+2.175075487,3.418238963
+2.208974156,3.418028038
+2.242872826,3.417817113
+2.276771495,3.417606188
+2.310670165,3.417395263
+2.344568834,3.417184338
+2.378467504,3.416973413
+2.412366173,3.416762488
+2.446264843,3.416551563
+2.480163512,3.416340638
+2.503649469,3.408526823
+2.509371154,3.408526823
+2.65,3.408526823
+2.75,3.408526823
+3,3.408526823
+4,3.408526823
+5,3.408526823
+7,3.408526823
+10,3.408526823
+20,3.408526823
+50,3.408526823
+100,3.408526823
+1000000,3.408526823
