+++ OWC_writeup/scripts/Eta_performance_table.csv
@@ -1,0 +1,117 @@
+inv_TSR,Eta
+-1000000,0.003429765
+-100,0.003429765
+-50,0.003429765
+-20,0.003429765
+-10,0.003429765
+-5,0.003429765
+-3,0.003429765
+-1,0.003429765
+-0.562,0.0065
+-0.17,0.011477158
+0.08,0.03
+0.263603728,0.075190746
+0.285361488,0.109436258
+0.303031426,0.133695731
+0.312196059,0.147823096
+0.315338846,0.155766371
+0.321228826,0.163040511
+0.32568587,0.173411038
+0.332745054,0.183500092
+0.337165679,0.191855048
+0.344625482,0.199981875
+0.356436838,0.217788488
+0.363274991,0.225356447
+0.36907706,0.233872485
+0.376260575,0.243477412
+0.382753366,0.252925278
+0.389798736,0.261402283
+0.39763153,0.269837529
+0.408171956,0.278543469
+0.417358566,0.287291074
+0.428092394,0.297463119
+0.437458592,0.305032595
+0.448874352,0.313937652
+0.459423569,0.323279961
+0.472478225,0.33256815
+0.485210544,0.341011385
+0.496411761,0.348811865
+0.511279564,0.357759094
+0.523608961,0.364826008
+0.538073843,0.372254637
+0.55573792,0.379637054
+0.571320621,0.386450734
+0.587069095,0.393406272
+0.608826855,0.400963161
+0.632035133,0.40850769
+0.658144445,0.414986945
+0.688605309,0.422387209
+0.720516691,0.42837799
+0.752428072,0.432588372
+0.784339454,0.436196157
+0.816250835,0.439858725
+0.848162217,0.442261318
+0.880073598,0.443842188
+0.91198498,0.445368277
+0.943896361,0.446593068
+0.975807743,0.446996136
+1.007719124,0.446577483
+1.039630506,0.446323173
+1.071541888,0.445082797
+1.103453269,0.444143719
+1.135364651,0.442136402
+1.167276032,0.440539946
+1.199187414,0.438258721
+1.231098795,0.435867934
+1.263010177,0.433093676
+1.294921558,0.430319418
+1.32683294,0.427736895
+1.358744321,0.42468873
+1.390655703,0.421996644
+1.422567085,0.419058041
+1.454478466,0.416420737
+1.486389848,0.413536916
+1.518301229,0.410680486
+1.550212611,0.408180135
+1.582123992,0.405679785
+1.614035374,0.402987699
+1.645946755,0.400213441
+1.677858137,0.397658309
+1.709769518,0.394938833
+1.7416809,0.392246747
+1.773592282,0.389664224
+1.805503663,0.386753012
+1.837415045,0.384225271
+1.869326426,0.381752311
+1.901237808,0.379251961
+1.933149189,0.376724219
+1.965060571,0.374196478
+1.996971952,0.371696128
+2.057893681,0.367054705
+2.089805062,0.36477348
+2.121716444,0.362245739
+2.153627825,0.359854951
+2.185539207,0.357546336
+2.217450589,0.352901715
+2.281273352,0.348065358
+2.345096115,0.343338564
+2.408918878,0.338748725
+2.469840606,0.33688156
+3.27,0.33688156
+5,0.33688156
+10,0.33688156
+20,0.33688156
+50,0.33688156
+100,0.33688156
+1000000,0.33688156
+2.65,3.408526823
+2.75,3.408526823
+3,3.408526823
+4,3.408526823
+5,3.408526823
+7,3.408526823
+10,3.408526823
+20,3.408526823
+50,3.408526823
+100,3.408526823
+1000000,3.408526823
