+++ OWC_writeup/scripts/archive/ClosedFormOWC.py
