+++ OWC_writeup/scripts/archive/example_implicitsolve_inside_WecOptTool.py
