+++ OWC_writeup/scripts/helper_functions/brent.py
